+++ jwp-basic-gradle/src/main/java/core/di/beans/factory/support/BeanFactory.java
@@ -1,8 +1,9 @@
-package core.di.factory;
+package core.di.beans.factory.support;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import core.annotation.PostConstruct;
+import core.di.context.annotation.AnnotatedBeanDefinition;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
