+++ jwp-basic-gradle/src/main/java/core/di/beans/factory/support/BeanFactoryUtils.java
@@ -1,4 +1,4 @@
-package core.di.factory;
+package core.di.beans.factory.support;
 
 import static org.reflections.ReflectionUtils.getAllConstructors;
 import static org.reflections.ReflectionUtils.getAllFields;
