+++ jwp-basic-gradle/src/main/java/core/di/beans/factory/support/InjectType.java
@@ -1,4 +1,4 @@
-package core.di.factory;
+package core.di.beans.factory.support;
 
 public enum InjectType {
     INJECT_NO, INJECT_CONSTRUCTOR, INJECT_FIELD
