+++ jwp-basic-gradle/src/main/java/core/di/context/support/AnnotationConfigApplicationContext.java
@@ -1,8 +1,11 @@
-package core.di.factory;
+package core.di.context.support;
 
 import com.google.common.collect.Lists;
 import core.annotation.ComponentScan;
-import core.di.factory.support.AnnotatedBeanDefinitionReader;
+import core.di.context.ApplicationContext;
+import core.di.beans.factory.support.BeanFactory;
+import core.di.context.annotation.AnnotatedBeanDefinitionReader;
+import core.di.context.annotation.ClasspathBeanDefinitionScanner;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Set;
