+++ jwp-basic-gradle/src/main/java/core/web/mvcframework/mapping/annotation/AnnotationHandlerMapping.java
@@ -4,7 +4,7 @@ import com.google.common.collect.Maps;
 import core.annotation.Controller;
 import core.annotation.RequestMapping;
 import core.annotation.RequestMethod;
-import core.di.factory.ApplicationContext;
+import core.di.context.ApplicationContext;
 import core.web.mvcframework.mapping.HandlerMapping;
 import java.lang.reflect.Method;
 import java.util.Map;
